# BDN config for running against local build of FSharp.Core.dll