BDN runs every benchmark for every job defined in the config

Every job can contain some custom settings, including configuration passed to MsBuild when building the auto-generated project.

So by using:

```fs
job.WithCustomBuildConfiguration("LocalBuild")
```

and

```xml
<ItemGroup Condition=" '$(Configuration)' == 'LocalBuild' ">
```

we can control the msbuild settings per BDN job

How to run the benchmarks:

```cmd
dotnet run -c Release
```

Sample results:

|    Method |    Job | BuildConfiguration | Size |      Mean |     Error |    StdDev | Ratio | Rank | Code Size | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |------- |------------------- |----- |----------:|----------:|----------:|------:|-----:|----------:|------:|------:|------:|----------:|
| optimized |  After |         LocalBuild |    1 | 0.2644 ns | 0.0022 ns | 0.0020 ns |  1.00 |    1 |     217 B |     - |     - |     - |         - |
|  original |  After |         LocalBuild |    1 | 0.3981 ns | 0.0039 ns | 0.0038 ns |  1.51 |    2 |     223 B |     - |     - |     - |         - |
|  nullOnly |  After |         LocalBuild |    1 | 0.2640 ns | 0.0024 ns | 0.0023 ns |  1.00 |    1 |     217 B |     - |     - |     - |         - |
| nullRefEq |  After |         LocalBuild |    1 | 0.2635 ns | 0.0023 ns | 0.0020 ns |  1.00 |    1 |     217 B |     - |     - |     - |         - |
|           |        |                    |      |           |           |           |       |      |           |       |       |       |           |
| optimized | Before |            Default |    1 | 0.3980 ns | 0.0038 ns | 0.0059 ns |  1.00 |    2 |     223 B |     - |     - |     - |         - |
|  original | Before |            Default |    1 | 0.3988 ns | 0.0039 ns | 0.0040 ns |  0.99 |    2 |     223 B |     - |     - |     - |         - |
|  nullOnly | Before |            Default |    1 | 0.2640 ns | 0.0021 ns | 0.0021 ns |  0.66 |    1 |     217 B |     - |     - |     - |         - |
| nullRefEq | Before |            Default |    1 | 0.2641 ns | 0.0014 ns | 0.0014 ns |  0.66 |    1 |     217 B |     - |     - |     - |         - |

Full log with detailed .dll version printed to console (to verify we are using the thing we think we are using):

<details>

```log
Assembly: "file:///C:/Projects/forks/StringLength-Optimization/StringLengthOptim/bin/Release/netcoreapp3.1/FSharp.Core.dll"
// Validating benchmarks:
// ***** BenchmarkRunner: Start   *****
// ***** Found 8 benchmark(s) in total *****
// ***** Building 2 exe(s) in Parallel: Start   *****
// ***** Done, took 00:00:17 (17.5 sec)   *****
// Found 8 benchmarks:
//   BenchLength.optimized: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.original: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.nullOnly: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.nullRefEq: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.optimized: Before(MaxRelativeError=0.01, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.original: Before(MaxRelativeError=0.01, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.nullOnly: Before(MaxRelativeError=0.01, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
//   BenchLength.nullRefEq: Before(MaxRelativeError=0.01, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]

// **************************
// Benchmark: BenchLength.optimized: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
// *** Execute ***
// Launch: 1 / 1
// Execute: dotnet "After.dll" --benchmarkName "FSharp.Perf.BenchLength.optimized(Size: 1)" --job "After" --benchmarkId 0 in C:\Projects\forks\StringLength-Optimization\StringLengthOptim\bin\Release\netcoreapp3.1\After\bin\LocalBuild\netcoreapp3.1
// BeforeAnythingElse

// Benchmark Process Environment Information:
// Runtime=.NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT DEBUG
// GC=Concurrent Workstation
// Job: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1)

Assembly: "file:///C:/Projects/forks/StringLength-Optimization/StringLengthOptim/bin/Release/netcoreapp3.1/After/bin/LocalBuild/netcoreapp3.1/FSharp.Core.dll"
Version: "42.42.42.42424"
OverheadJitting  1: 10000 op, 271700.00 ns, 27.1700 ns/op
WorkloadJitting  1: 10000 op, 365100.00 ns, 36.5100 ns/op

OverheadJitting  2: 160000 op, 179500.00 ns, 1.1219 ns/op
WorkloadJitting  2: 160000 op, 213400.00 ns, 1.3337 ns/op

WorkloadPilot    1: 160000 op, 45700.00 ns, 0.2856 ns/op
WorkloadPilot    2: 875360000 op, 232730700.00 ns, 0.2659 ns/op
WorkloadPilot    3: 940320000 op, 250452300.00 ns, 0.2663 ns/op
WorkloadPilot    4: 938720000 op, 250016200.00 ns, 0.2663 ns/op

OverheadWarmup   1: 938720000 op, 210200.00 ns, 0.0002 ns/op
OverheadWarmup   2: 938720000 op, 207900.00 ns, 0.0002 ns/op
OverheadWarmup   3: 938720000 op, 206400.00 ns, 0.0002 ns/op
OverheadWarmup   4: 938720000 op, 206500.00 ns, 0.0002 ns/op
OverheadWarmup   5: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadWarmup   6: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadWarmup   7: 938720000 op, 206300.00 ns, 0.0002 ns/op

OverheadActual   1: 938720000 op, 207900.00 ns, 0.0002 ns/op
OverheadActual   2: 938720000 op, 206400.00 ns, 0.0002 ns/op
OverheadActual   3: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual   4: 938720000 op, 207800.00 ns, 0.0002 ns/op
OverheadActual   5: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual   6: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual   7: 938720000 op, 207500.00 ns, 0.0002 ns/op
OverheadActual   8: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual   9: 938720000 op, 206200.00 ns, 0.0002 ns/op
OverheadActual  10: 938720000 op, 204700.00 ns, 0.0002 ns/op
OverheadActual  11: 938720000 op, 206400.00 ns, 0.0002 ns/op
OverheadActual  12: 938720000 op, 204700.00 ns, 0.0002 ns/op
OverheadActual  13: 938720000 op, 206200.00 ns, 0.0002 ns/op
OverheadActual  14: 938720000 op, 206200.00 ns, 0.0002 ns/op
OverheadActual  15: 938720000 op, 206100.00 ns, 0.0002 ns/op

WorkloadWarmup   1: 938720000 op, 250269400.00 ns, 0.2666 ns/op

// BeforeActualRun
WorkloadActual   1: 938720000 op, 250153800.00 ns, 0.2665 ns/op
WorkloadActual   2: 938720000 op, 250554800.00 ns, 0.2669 ns/op
WorkloadActual   3: 938720000 op, 249943200.00 ns, 0.2663 ns/op
WorkloadActual   4: 938720000 op, 250232100.00 ns, 0.2666 ns/op
WorkloadActual   5: 938720000 op, 250411800.00 ns, 0.2668 ns/op
WorkloadActual   6: 938720000 op, 249978700.00 ns, 0.2663 ns/op
WorkloadActual   7: 938720000 op, 251001900.00 ns, 0.2674 ns/op
WorkloadActual   8: 938720000 op, 247116800.00 ns, 0.2632 ns/op
WorkloadActual   9: 938720000 op, 246125400.00 ns, 0.2622 ns/op
WorkloadActual  10: 938720000 op, 246357400.00 ns, 0.2624 ns/op
WorkloadActual  11: 938720000 op, 247624000.00 ns, 0.2638 ns/op
WorkloadActual  12: 938720000 op, 246929300.00 ns, 0.2630 ns/op
WorkloadActual  13: 938720000 op, 247430500.00 ns, 0.2636 ns/op
WorkloadActual  14: 938720000 op, 246966000.00 ns, 0.2631 ns/op
WorkloadActual  15: 938720000 op, 245794900.00 ns, 0.2618 ns/op

// AfterActualRun
WorkloadResult   1: 938720000 op, 249947500.00 ns, 0.2663 ns/op
WorkloadResult   2: 938720000 op, 250348500.00 ns, 0.2667 ns/op
WorkloadResult   3: 938720000 op, 249736900.00 ns, 0.2660 ns/op
WorkloadResult   4: 938720000 op, 250025800.00 ns, 0.2663 ns/op
WorkloadResult   5: 938720000 op, 250205500.00 ns, 0.2665 ns/op
WorkloadResult   6: 938720000 op, 249772400.00 ns, 0.2661 ns/op
WorkloadResult   7: 938720000 op, 250795600.00 ns, 0.2672 ns/op
WorkloadResult   8: 938720000 op, 246910500.00 ns, 0.2630 ns/op
WorkloadResult   9: 938720000 op, 245919100.00 ns, 0.2620 ns/op
WorkloadResult  10: 938720000 op, 246151100.00 ns, 0.2622 ns/op
WorkloadResult  11: 938720000 op, 247417700.00 ns, 0.2636 ns/op
WorkloadResult  12: 938720000 op, 246723000.00 ns, 0.2628 ns/op
WorkloadResult  13: 938720000 op, 247224200.00 ns, 0.2634 ns/op
WorkloadResult  14: 938720000 op, 246759700.00 ns, 0.2629 ns/op
WorkloadResult  15: 938720000 op, 245588600.00 ns, 0.2616 ns/op
GC:  0 0 0 480 938720000
Threading:  2 0 938720000

// AfterAll
// Benchmark Process 14196 has exited with code 0

Mean = 0.264 ns, StdErr = 0.001 ns (0.20%), N = 15, StdDev = 0.002 ns
Min = 0.262 ns, Q1 = 0.263 ns, Median = 0.264 ns, Q3 = 0.266 ns, Max = 0.267 ns
IQR = 0.003 ns, LowerFence = 0.258 ns, UpperFence = 0.271 ns
ConfidenceInterval = [0.262 ns; 0.267 ns] (CI 99.9%), Margin = 0.002 ns (0.82% of Mean)
Skewness = 0.01, Kurtosis = 1.13, MValue = 2

// **************************
// Benchmark: BenchLength.original: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
// *** Execute ***
// Launch: 1 / 1
// Execute: dotnet "After.dll" --benchmarkName "FSharp.Perf.BenchLength.original(Size: 1)" --job "After" --benchmarkId 1 in C:\Projects\forks\StringLength-Optimization\StringLengthOptim\bin\Release\netcoreapp3.1\After\bin\LocalBuild\netcoreapp3.1
// BeforeAnythingElse

// Benchmark Process Environment Information:
// Runtime=.NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT DEBUG
// GC=Concurrent Workstation
// Job: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1)

Assembly: "file:///C:/Projects/forks/StringLength-Optimization/StringLengthOptim/bin/Release/netcoreapp3.1/After/bin/LocalBuild/netcoreapp3.1/FSharp.Core.dll"
Version: "42.42.42.42424"
OverheadJitting  1: 10000 op, 292100.00 ns, 29.2100 ns/op
WorkloadJitting  1: 10000 op, 455200.00 ns, 45.5200 ns/op

OverheadJitting  2: 160000 op, 171500.00 ns, 1.0719 ns/op
WorkloadJitting  2: 160000 op, 236600.00 ns, 1.4788 ns/op

WorkloadPilot    1: 160000 op, 89500.00 ns, 0.5594 ns/op
WorkloadPilot    2: 447040000 op, 178664900.00 ns, 0.3997 ns/op
WorkloadPilot    3: 625600000 op, 249599900.00 ns, 0.3990 ns/op
WorkloadPilot    4: 626720000 op, 254990600.00 ns, 0.4069 ns/op
WorkloadPilot    5: 614560000 op, 241683000.00 ns, 0.3933 ns/op
WorkloadPilot    6: 635840000 op, 250878700.00 ns, 0.3946 ns/op
WorkloadPilot    7: 633760000 op, 255017800.00 ns, 0.4024 ns/op

OverheadWarmup   1: 633760000 op, 143800.00 ns, 0.0002 ns/op
OverheadWarmup   2: 633760000 op, 140600.00 ns, 0.0002 ns/op
OverheadWarmup   3: 633760000 op, 139600.00 ns, 0.0002 ns/op
OverheadWarmup   4: 633760000 op, 138300.00 ns, 0.0002 ns/op
OverheadWarmup   5: 633760000 op, 139400.00 ns, 0.0002 ns/op
OverheadWarmup   6: 633760000 op, 140700.00 ns, 0.0002 ns/op
OverheadWarmup   7: 633760000 op, 138300.00 ns, 0.0002 ns/op
OverheadWarmup   8: 633760000 op, 138400.00 ns, 0.0002 ns/op
OverheadWarmup   9: 633760000 op, 139500.00 ns, 0.0002 ns/op
OverheadWarmup  10: 633760000 op, 138400.00 ns, 0.0002 ns/op

OverheadActual   1: 633760000 op, 139400.00 ns, 0.0002 ns/op
OverheadActual   2: 633760000 op, 140500.00 ns, 0.0002 ns/op
OverheadActual   3: 633760000 op, 293200.00 ns, 0.0005 ns/op
OverheadActual   4: 633760000 op, 141300.00 ns, 0.0002 ns/op
OverheadActual   5: 633760000 op, 139500.00 ns, 0.0002 ns/op
OverheadActual   6: 633760000 op, 139300.00 ns, 0.0002 ns/op
OverheadActual   7: 633760000 op, 139300.00 ns, 0.0002 ns/op
OverheadActual   8: 633760000 op, 138400.00 ns, 0.0002 ns/op
OverheadActual   9: 633760000 op, 139400.00 ns, 0.0002 ns/op
OverheadActual  10: 633760000 op, 139300.00 ns, 0.0002 ns/op
OverheadActual  11: 633760000 op, 140100.00 ns, 0.0002 ns/op
OverheadActual  12: 633760000 op, 139400.00 ns, 0.0002 ns/op
OverheadActual  13: 633760000 op, 140400.00 ns, 0.0002 ns/op
OverheadActual  14: 633760000 op, 139500.00 ns, 0.0002 ns/op
OverheadActual  15: 633760000 op, 138400.00 ns, 0.0002 ns/op

WorkloadWarmup   1: 633760000 op, 253651100.00 ns, 0.4002 ns/op

// BeforeActualRun
WorkloadActual   1: 633760000 op, 252714100.00 ns, 0.3988 ns/op
WorkloadActual   2: 633760000 op, 252622000.00 ns, 0.3986 ns/op
WorkloadActual   3: 633760000 op, 252755300.00 ns, 0.3988 ns/op
WorkloadActual   4: 633760000 op, 253550100.00 ns, 0.4001 ns/op
WorkloadActual   5: 633760000 op, 256465100.00 ns, 0.4047 ns/op
WorkloadActual   6: 633760000 op, 256179700.00 ns, 0.4042 ns/op
WorkloadActual   7: 633760000 op, 255008300.00 ns, 0.4024 ns/op
WorkloadActual   8: 633760000 op, 264614600.00 ns, 0.4175 ns/op
WorkloadActual   9: 633760000 op, 255411700.00 ns, 0.4030 ns/op
WorkloadActual  10: 633760000 op, 252671900.00 ns, 0.3987 ns/op
WorkloadActual  11: 633760000 op, 251819200.00 ns, 0.3973 ns/op
WorkloadActual  12: 633760000 op, 249656800.00 ns, 0.3939 ns/op
WorkloadActual  13: 633760000 op, 250236000.00 ns, 0.3948 ns/op
WorkloadActual  14: 633760000 op, 249989700.00 ns, 0.3945 ns/op
WorkloadActual  15: 633760000 op, 249760200.00 ns, 0.3941 ns/op
WorkloadActual  16: 633760000 op, 248996800.00 ns, 0.3929 ns/op
WorkloadActual  17: 633760000 op, 251083000.00 ns, 0.3962 ns/op

// AfterActualRun
WorkloadResult   1: 633760000 op, 252574700.00 ns, 0.3985 ns/op
WorkloadResult   2: 633760000 op, 252482600.00 ns, 0.3984 ns/op
WorkloadResult   3: 633760000 op, 252615900.00 ns, 0.3986 ns/op
WorkloadResult   4: 633760000 op, 253410700.00 ns, 0.3999 ns/op
WorkloadResult   5: 633760000 op, 256325700.00 ns, 0.4045 ns/op
WorkloadResult   6: 633760000 op, 256040300.00 ns, 0.4040 ns/op
WorkloadResult   7: 633760000 op, 254868900.00 ns, 0.4022 ns/op
WorkloadResult   8: 633760000 op, 255272300.00 ns, 0.4028 ns/op
WorkloadResult   9: 633760000 op, 252532500.00 ns, 0.3985 ns/op
WorkloadResult  10: 633760000 op, 251679800.00 ns, 0.3971 ns/op
WorkloadResult  11: 633760000 op, 249517400.00 ns, 0.3937 ns/op
WorkloadResult  12: 633760000 op, 250096600.00 ns, 0.3946 ns/op
WorkloadResult  13: 633760000 op, 249850300.00 ns, 0.3942 ns/op
WorkloadResult  14: 633760000 op, 249620800.00 ns, 0.3939 ns/op
WorkloadResult  15: 633760000 op, 248857400.00 ns, 0.3927 ns/op
WorkloadResult  16: 633760000 op, 250943600.00 ns, 0.3960 ns/op
GC:  0 0 0 5528 633760000
Threading:  2 0 633760000

// AfterAll
// Benchmark Process 2352 has exited with code 0

Mean = 0.398 ns, StdErr = 0.001 ns (0.24%), N = 16, StdDev = 0.004 ns
Min = 0.393 ns, Q1 = 0.395 ns, Median = 0.398 ns, Q3 = 0.400 ns, Max = 0.404 ns
IQR = 0.006 ns, LowerFence = 0.386 ns, UpperFence = 0.409 ns
ConfidenceInterval = [0.394 ns; 0.402 ns] (CI 99.9%), Margin = 0.004 ns (0.97% of Mean)
Skewness = 0.24, Kurtosis = 1.67, MValue = 2

// **************************
// Benchmark: BenchLength.nullOnly: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1) [Size=1]
// *** Execute ***
// Launch: 1 / 1
// Execute: dotnet "After.dll" --benchmarkName "FSharp.Perf.BenchLength.nullOnly(Size: 1)" --job "After" --benchmarkId 2 in C:\Projects\forks\StringLength-Optimization\StringLengthOptim\bin\Release\netcoreapp3.1\After\bin\LocalBuild\netcoreapp3.1
// BeforeAnythingElse

// Benchmark Process Environment Information:
// Runtime=.NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT DEBUG
// GC=Concurrent Workstation
// Job: After(MaxRelativeError=0.01, BuildConfiguration=LocalBuild, IterationTime=250.0000 ms, WarmupCount=1)

Assembly: "file:///C:/Projects/forks/StringLength-Optimization/StringLengthOptim/bin/Release/netcoreapp3.1/After/bin/LocalBuild/netcoreapp3.1/FSharp.Core.dll"
Version: "42.42.42.42424"
OverheadJitting  1: 10000 op, 293400.00 ns, 29.3400 ns/op
WorkloadJitting  1: 10000 op, 383900.00 ns, 38.3900 ns/op

OverheadJitting  2: 160000 op, 162300.00 ns, 1.0144 ns/op
WorkloadJitting  2: 160000 op, 203500.00 ns, 1.2719 ns/op

WorkloadPilot    1: 160000 op, 45500.00 ns, 0.2844 ns/op
WorkloadPilot    2: 879200000 op, 234308300.00 ns, 0.2665 ns/op
WorkloadPilot    3: 938080000 op, 249855700.00 ns, 0.2663 ns/op
WorkloadPilot    4: 938720000 op, 250001800.00 ns, 0.2663 ns/op

OverheadWarmup   1: 938720000 op, 210900.00 ns, 0.0002 ns/op
OverheadWarmup   2: 938720000 op, 208100.00 ns, 0.0002 ns/op
OverheadWarmup   3: 938720000 op, 207900.00 ns, 0.0002 ns/op
OverheadWarmup   4: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadWarmup   5: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadWarmup   6: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadWarmup   7: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadWarmup   8: 938720000 op, 204700.00 ns, 0.0002 ns/op

OverheadActual   1: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual   2: 938720000 op, 206400.00 ns, 0.0002 ns/op
OverheadActual   3: 938720000 op, 206500.00 ns, 0.0002 ns/op
OverheadActual   4: 938720000 op, 206500.00 ns, 0.0002 ns/op
OverheadActual   5: 938720000 op, 207900.00 ns, 0.0002 ns/op
OverheadActual   6: 938720000 op, 206200.00 ns, 0.0002 ns/op
OverheadActual   7: 938720000 op, 206200.00 ns, 0.0002 ns/op
OverheadActual   8: 938720000 op, 208400.00 ns, 0.0002 ns/op
OverheadActual   9: 938720000 op, 371700.00 ns, 0.0004 ns/op
OverheadActual  10: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual  11: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual  12: 938720000 op, 207100.00 ns, 0.0002 ns/op
OverheadActual  13: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual  14: 938720000 op, 206300.00 ns, 0.0002 ns/op
OverheadActual  15: 938720000 op, 206200.00 ns, 0.0002 ns/op

WorkloadWarmup   1: 938720000 op, 250379100.00 ns, 0.2667 ns/op

// BeforeActualRun
WorkloadActual   1: 938720000 op, 250468200.00 ns, 0.2668 ns/op
WorkloadActual   2: 938720000 op, 250099500.00 ns, 0.2664 ns/op
WorkloadActual   3: 938720000 op, 250112200.00 ns, 0.2664 ns/op
WorkloadActual   4: 938720000 op, 250194700.00 ns, 0.2665 ns/op
WorkloadActual   5: 938720000 op, 249650100.00 ns, 0.2659 ns/op
WorkloadActual   6: 938720000 op, 250067700.00 ns, 0.2664 ns/op
WorkloadActual   7: 938720000 op, 249990400.00 ns, 0.2663 ns/op
WorkloadActual   8: 938720000 op, 246975300.00 ns, 0.2631 ns/op
WorkloadActual   9: 938720000 op, 248032900.00 ns, 0.2642 ns/op
WorkloadActual  10: 938720000 op, 245958500.00 ns, 0.2620 ns/op
WorkloadActual  11: 938720000 op, 245901100.00 ns, 0.2620 ns/op
WorkloadActual  12: 938720000 op, 244811600.00 ns, 0.2608 ns/op
WorkloadActual  13: 938720000 op, 245347800.00 ns, 0.2614 ns/op
WorkloadActual  14: 938720000 op, 245987200.00 ns, 0.2620 ns/op
WorkloadActual  15: 938720000 op, 246974100.00 ns, 0.2631 ns/op

// AfterActualRun
WorkloadResult   1: 938720000 op, 250261900.00 ns, 0.2666 ns/op
WorkloadResult   2: 938720000 op, 249893200.00 ns, 0.2662 ns/op
WorkloadResult   3: 938720000 op, 249905900.00 ns, 0.2662 ns/op
WorkloadResult   4: 938720000 op, 249988400.00 ns, 0.2663 ns/op
WorkloadResult   5: 938720000 op, 249443800.00 ns, 0.2657 ns/op
WorkloadResult   6: 938720000 op, 249861400.00 ns, 0.2662 ns/op
WorkloadResult   7: 938720000 op, 249784100.00 ns, 0.2661 ns/op
WorkloadResult   8: 938720000 op, 246769000.00 ns, 0.2629 ns/op
WorkloadResult   9: 938720000 op, 247826600.00 ns, 0.2640 ns/op
WorkloadResult  10: 938720000 op, 245752200.00 ns, 0.2618 ns/op
WorkloadResult  11: 938720000 op, 245694800.00 ns, 0.2617 ns/op
WorkloadResult  12: 938720000 op, 244605300.00 ns, 0.2606 ns/op
WorkloadResult  13: 938720000 op, 245141500.00 ns, 0.2611 ns/op
WorkloadResult  14: 938720000 op, 245780900.00 ns, 0.2618 ns/op
WorkloadResult  15: 938720000 op, 246767800.00 ns, 0.2629 ns/op
GC:  0 0 0 0 938720000
Threading:  2 0 938720000

// AfterAll
…