@abelbraaksma 2020-06-26
This looks very interesting, thank you!!!

Let me see if I understand this approach correctly:

1. Make sure you add `<Optimize>true</Optimize>`, because we'll add a new config and otherwise we'll end up with and unoptimized build
2. Add a new build config, say "LocalBuild", and use the MSBuild syntax to specifically enable the wanted lines in that case
3. BDN will always do a "Default" build with default settings
4. Use the "fluent" syntax to config the benchmark, as this option cannot be set with just attributes
5. Create a `Job` for each configuration, the special one with `WithCustomBuildConfiguration`
6. Add those jobs through `AddJob` to the default config

I'm amazed that this is possible, this gives so many possibilities for my particular scenarios!

It'll also help to create "before" and "after" graphs automatically in the CI system.