@adamsitnik 2020-06-26
> Let me see if I understand this approach correctly

I confirm that your understanding is correct.

> I'm amazed that this is possible

I am glad that you like it!